Collection point master create update and search end point changes.

+++ swm/swm-services/src/main/java/org/egov/swm/domain/model/Vehicle.java
@@ -48,7 +48,7 @@ public class Vehicle {
 
 	@NotNull
 	@JsonProperty("vehicleCapacity")
-	private Long vehicleCapacity = null;
+	private Double vehicleCapacity = null;
 
 	@NotNull
 	@JsonProperty("numberOfPersonsReq")
